backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9
pymupdf>=1.24.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
tavily-python>=0.3.0
groq>=0.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
